Struct dao::bid_escrow::bid_engine::BidEngine

pub struct BidEngine { /* private fields */ }

Manages the Bidding process.

Implementations

impl BidEngine

pub fn job_offers_count(&self) -> u32

Gets the total number of JobOffers.

pub fn bids_count(&self) -> u32

Gets the total number of Bids.

pub fn get_job_offer(&self, job_offer_id: JobOfferId) -> Option<JobOffer>

Gets the JobOffer with a given id or None.

pub fn get_job_offer_or_revert(&self, job_offer_id: &JobOfferId) -> JobOffer

Gets the JobOffer with a given id or reverts with JobOfferNotFound.

pub fn get_bid(&self, bid_id: BidId) -> Option<Bid>

Gets the Bid with a given id or None.

pub fn get_bid_or_revert(&self, bid_id: &BidId) -> Bid

Gets the Bid with a given id or reverts with BidNotFound.

pub fn next_bid_id(&mut self) -> BidId

Increments bid counter.

pub fn store_bid(&mut self, bid: Bid)

Writes the Bid to the storage.

pub fn get_job_offer_configuration(&self, job: &Job) -> Configuration

Gets the Configuration of the Job.

pub fn post_job_offer( &mut self, expected_timeframe: BlockTime, budget: Balance, dos_fee: Balance )

pub fn submit_bid( &mut self, job_offer_id: JobOfferId, time: BlockTime, payment: Balance, reputation_stake: Balance, onboard: bool, cspr_stake: Option<Balance> )

pub fn cancel_bid(&mut self, bid_id: BidId)

pub fn cancel_job_offer(&mut self, job_offer_id: JobOfferId)

Invalidates the Job Offer, returns DOS Fee to the Job Poster, returns funds to Bidders.

If a Job with the given id does not exists, contract execution stop with Error::JobOfferNotFound.

Executes validations: HasPermissionsToCancelJobOffer and CanJobOfferBeCancelled.

pub fn pick_bid( &mut self, job_offer_id: JobOfferId, bid_id: BidId, cspr_amount: Balance )

pub fn slash_voter(&mut self, voter: Address) -> (Vec<JobOfferId>, Vec<BidId>)

impl BidEngine

pub fn cancel_all_bids(&mut self, job_offer_id: &JobOfferId)

pub fn return_job_offer_poster_dos_fee(&mut self, job_offer_id: &JobOfferId)