Struct dao::bid_escrow::job_engine::JobEngine

pub struct JobEngine { /* private fields */ }

Manages Jobs lifecycle.

Implementations

impl JobEngine

pub fn submit_job_proof(&mut self, job_id: JobId, proof: DocumentHash)

Validates the correctness of proof submission. If the submission is correct, the Job Storage is updated, the Voting process starts.

Errors

If a proof has been submitted before, reverts with Error::JobAlreadySubmitted.

pub fn submit_job_proof_during_grace_period( &mut self, job_id: JobId, proof: DocumentHash, reputation_stake: Balance, onboard: bool )

Updates the old Bid and Job, the job is assigned to a new Worker. The rest goes the same as regular proof submission. See submit_job_proof(). The old Worker who didn’t submit the proof in time, is getting slashed.

See the Grace Period section in the module description.

pub fn cancel_job(&mut self, job_id: JobId)

Terminates the Voting process and slashes the Worker.

the bid stake is redistributed along the VAs’ and the multisig wallet.

DOS Fee is returned to the Job Poster.

Error

If the state in which the process cannot be canceled, the execution reverts with Error::CannotCancelJob or Error::JobCannotBeYetCanceled.

pub fn vote( &mut self, voting_id: VotingId, voting_type: VotingType, choice: Choice, stake: Balance )

Records vote in Voting.

Error

Error::CannotVoteOnOwnJob.

pub fn finish_voting( &mut self, voting_id: VotingId, voting_type: VotingType ) -> VotingSummary

Ends the current voting phase and redistributes funds.

Interacts with Reputation Token Contract to redistribute reputation.

pub fn slash_voter( &mut self, voter: Address ) -> (Vec<JobId>, Vec<VotingId>, Vec<VotingId>)

Slashes the voter and cancels all his active jobs. Returns the lists of slashed jobs, canceled votings and affected votings.

pub fn jobs_count(&self) -> u32

Returns the total number of jobs.

pub fn get_job(&self, job_id: JobId) -> Option<Job>

Gets the job with a given id or None.

impl JobEngine

pub fn return_job_poster_payment_and_dos_fee(&mut self, job: &Job)

pub fn return_external_worker_cspr_stake(&mut self, job: &Job)