Struct dao::bid_escrow::storage::JobStorage

pub struct JobStorage { /* private fields */ }

Stores Job-related variables and mappings.

Implementations

impl JobStorage

pub fn store_job_for_voting(&mut self, voting_id: VotingId, job_id: JobId)

Links voting with a job.

pub fn get_job(&self, job_id: JobId) -> Option<Job>

Gets the Job with a given id or None.

pub fn get_job_by_voting_id(&self, voting_id: VotingId) -> Job

Gets the Job matching to a given id or reverts with VotingIdNotFound.

pub fn get_job_or_revert(&self, job_id: JobId) -> Job

Gets the Job with a given id or reverts with MappingItemNotAvailable.

pub fn store_job(&mut self, job: Job)

Writes a Job to the storage.

pub fn jobs_count(&self) -> u32

Gets the current value of jobs counter.

pub fn next_job_id(&mut self) -> JobId

Increments jobs counter.

pub fn add_to_active_jobs(&mut self, job_id: JobId)

Adds a job to the list of active jobs.

pub fn remove_from_active_jobs(&mut self, job_id: JobId)

Removes a job from the list of active jobs.

pub fn get_active_jobs(&self) -> Vec<JobId>

Returns all active jobs.