Struct dao::core_contracts::KycNftContract

pub struct KycNftContract { /* private fields */ }

NFT contract that tells the system if user is KYC’d. Kyc Owned Nft contract acts like an erc-721 token and derives most of erc-721 standard.

Kyc Owned Nft token is mintable and burnable but the caller needs to have permissions to perform those actions.

Implementations

impl KycNftContract

pub fn init(&mut self, name: String, symbol: String, base_uri: TokenUri)

Contract constructor.

Initializes DaoNft module.

pub fn propose_new_owner(&mut self, owner: Address)

Changes the ownership of the contract. Transfers ownership to the owner. Only the current owner is permitted to call this method. Read more

pub fn accept_new_owner(&mut self)

Accepts the new owner proposition. This can be called only by the proposed owner. Read more

pub fn add_to_whitelist(&mut self, address: Address)

Adds a new address to the whitelist.

pub fn remove_from_whitelist(&mut self, address: Address)

Remove address from the whitelist.

pub fn is_whitelisted(&self, address: Address) -> bool

Checks whether the given address is added to the whitelist.

pub fn get_owner(&self) -> Option<Address>

Returns the address of the current owner.

pub fn name(&self) -> String

Returns a descriptive name for a collection of tokens in this contract.

pub fn symbol(&self) -> String

Gets an abbreviated name for tokens in this contract.

pub fn owner_of(&self, token_id: &TokenId) -> Address

Returns the address of the owner of the token.

If the given token_id does not exist the None value is returned.

pub fn token_id(&self, address: Address) -> Option<TokenId>

Returns the token id for a given address.

If the owner does not own any token the None value is returned.

pub fn balance_of(&self, owner: &Address) -> U256

Returns the number of tokens owned by owner.

pub fn total_supply(&self) -> Balance

Returns the total number of tokens.

pub fn token_uri(&self, token_id: TokenId) -> TokenUri

Returns a distinct Uniform Resource Identifier (URI) for a given asset.

pub fn base_uri(&self) -> TokenUri

Returns a URI prefix that is used by all the assets.

pub fn mint(&mut self, to: Address)

Creates a new token with the next id and transfers it to a new owner. Increments the total supply and the balance of the to address.

Note

Only whitelisted addresses are permitted to call this function.

Each user is entitled to own only one token.

Errors

UserAlreadyOwnsToken if the to address already owns a token.

Events

Transfer event when minted successfully.

pub fn burn(&mut self, owner: Address)

Burns a token with the given id. Decrements the balance of the token owner and decrements the total supply.

Errors

NotWhitelisted if the caller is not whitelisted.

Events

Transfer event when burnt successfully.