Enum dao::io::DaoAction

pub enum DaoAction {
    AddToWhiteList(ActorId),
    SubmitMembershipProposal {
        applicant: ActorId,
        token_tribute: u128,
        shares_requested: u128,
        quorum: u128,
        details: String,
    },
    SubmitFundingProposal {
        applicant: ActorId,
        amount: u128,
        quorum: u128,
        details: String,
    },
    ProcessProposal(u128),
    SubmitVote {
        proposal_id: u128,
        vote: Vote,
    },
    RageQuit(u128),
    Abort(u128),
    UpdateDelegateKey(ActorId),
    SetAdmin(ActorId),
    Continue(u64),
}

Variants

AddToWhiteList(ActorId)

Tuple Fields

0: ActorId

valid actor ID

Adds members to whitelist.

Requirements:

Only admin can add actors to whitelist;

Member ID cant be zero;

Member can not be added to whitelist more than once;

On success replies with DaoEvent::MemberAddedToWhitelist

SubmitMembershipProposal

Fields

applicant: ActorId

an actor who wishes to become a DAO member

token_tribute: u128

the number of tokens the applicant offered for shares in DAO

shares_requested: u128

the amount of shares the applicant is requesting for his token tribute

quorum: u128

a certain threshold of YES votes in order for the proposal to pass

details: String

the proposal description

The proposal of joining the DAO.

Requirements:

The proposal can be submitted only by the existing members or their delegate addresses;

The applicant account must be either a DAO member or is in the whitelist.

On success replies with DaoEvent::SubmitMembershipProposal

SubmitFundingProposal

Fields

applicant: ActorId

an actor that will be funded

amount: u128

the number of fungible tokens that will be sent to the receiver

quorum: u128

a certain threshold of YES votes in order for the proposal to pass

details: String

the proposal description

The proposal of funding.

Requirements:

The proposal can be submitted only by the existing members or their delegate addresses;

The receiver ID can’t be the zero;

The DAO must have enough funds to finance the proposal;

On success replies with DaoEvent::SubmitFundingProposal

ProcessProposal(u128)

Tuple Fields

0: u128

the proposal ID

The proposal processing after the proposal completes during the grace period. If the membership proposal is accepted, the tribute tokens are deposited into the contract and new shares are minted and issued to the applicant. If the membership proposal is rejected, the tribute tokens are returned to the applicant. If the funding proposal is accepted, the indicated amount of tokens is transfered to the applicant; If the funging proposal is rejected, the indicated amount of tokens remains in the contract.

Requirements:

The previous proposal must be processed;

The proposal must exist and be ready for processing;

The proposal must not be aborted or already be processed.

On success replies with DaoEvent::ProcessProposal

SubmitVote

Fields

proposal_id: u128

the proposal ID

vote: Vote

the member a member vote (YES or NO)

The member (or the delegate address of the member) submits his vote (YES or NO) on the proposal.

Requirements:

The proposal can be submitted only by the existing members or their delegate addresses;

The member can vote on the proposal only once;

Proposal must exist, the voting period must has started and not expired;

Proposal must not be aborted.

On success replies with DaoEvent::SubmitVote

RageQuit(u128)

Tuple Fields

0: u128

The amount of shares the member would like to withdraw (the shares are converted to fungible tokens)

Withdraws the capital of the member.

Requirements:

msg::source() must be DAO member;

The member must have sufficient amount;

The latest proposal the member voted YES must be processed;

Admin can ragequit only after transferring his role to another actor.

On success replies with DaoEvent::RageQuit

Abort(u128)

Tuple Fields

0: u128

the proposal ID

Aborts the membership proposal. It can be used in case when applicant is disagree with the requested shares or the details the proposer indicated by the proposer.

Requirements:

msg::source() must be the applicant;

The proposal must be membership proposal;

The proposal can be aborted during only the abort window

The proposal has not be already aborted.

On success replies with DaoEvent::Abort

UpdateDelegateKey(ActorId)

Tuple Fields

0: ActorId

New delegate account

Sets the delegate key that is responsible for submitting proposals and voting; The deleagate key defaults to member address unless updated.

Requirements:

msg::source() must be DAO member;

The delegate key must not be zero address;

A delegate key can be assigned only to one member.

On success replies with DaoEvent::DelegateKeyUpdated

SetAdmin(ActorId)

Tuple Fields

0: ActorId

New admin account

Assigns the admin position to new actor.

Requirements:

Only admin can assign new admin.

On success replies with DaoEvent::AdminUpdated

Continue(u64)

Tuple Fields

0: u64

the transaction ID

Continues the transaction if it fails due to lack of gas or due to an error in the token contract.

Requirements:

Transaction must exist.

On success replies with the payload of continued transaction.