Struct dao::voting_contracts::OnboardingRequestContractRef

pub struct OnboardingRequestContractRef { /* private fields */ }

Reference to the OnboardingRequestContract contract instance.

Implementations

impl OnboardingRequestContractRef

pub fn at(address: &Address) -> Self

pub fn address(&self) -> &Address

pub fn with_tokens<T>(&self, amount: T) -> Self
where T: Into<Balance>,

impl OnboardingRequestContractRef

pub fn create_voting(&mut self, reason: DocumentHash)

Submits an onboarding request. If the request is valid voting starts.

Events

OnboardingVotingCreated

pub fn get_cspr_balance(&self) -> Balance

Gets the CSPR balance of the contract.

pub fn slash_voter(&mut self, voter: Address) -> SlashedVotings

Erases the voter from voting with the given id. Read more.

pub fn cancel_finished_voting(&mut self, voting_id: VotingId)

Cancels a voting that has not been finished in defined time

Errors

crate::utils::Error::VotingDoesNotExist

crate::utils::Error::VotingWithGivenTypeNotInProgress

crate::utils::Error::FinishingCompletedVotingNotAllowed

crate::utils::Error::VotingNotStarted

crate::utils::Error::VotingAlreadyFinished

[crate::utils::Error::VotingAlreadyCancelled]

pub fn voting_exists( &self, voting_id: VotingId, voting_type: VotingType ) -> bool

Checks if voting of a given type and id exists.

pub fn get_ballot( &self, voting_id: VotingId, voting_type: VotingType, address: Address ) -> Option<Ballot>

Returns the Voter’s Ballot.

pub fn get_voter( &self, voting_id: VotingId, voting_type: VotingType, at: u32 ) -> Option<Address>

Gets the address of nth voter who voted on Voting with voting_id.

pub fn get_voting(&self, voting_id: VotingId) -> Option<VotingStateMachine>

Returns Voting for given id.

pub fn propose_new_owner(&mut self, owner: Address)

Changes the ownership of the contract. Transfers ownership to the owner. Only the current owner is permitted to call this method. Read more

pub fn accept_new_owner(&mut self)

Accepts the new owner proposition. This can be called only by the proposed owner. Read more

pub fn add_to_whitelist(&mut self, address: Address)

Adds a new address to the whitelist. Read more

pub fn remove_from_whitelist(&mut self, address: Address)

Remove address from the whitelist. Read more

pub fn is_whitelisted(&self, address: Address) -> bool

Checks whether the given address is added to the whitelist. Read more.

pub fn get_owner(&self) -> Option<Address>

Returns the address of the current owner. Read more.

pub fn finish_voting( &mut self, voting_id: VotingId, voting_type: VotingType ) -> VotingSummary

Finishes voting stage. Depending on stage, the voting can be converted to a formal one, end with a refund or convert the requester to a VA.

pub fn vote( &mut self, voting_id: VotingId, voting_type: VotingType, choice: Choice, stake: Balance )

Casts a vote. Read more

pub fn variable_repository_address(&self) -> Address

Returns the address of Variable Repository contract.

pub fn reputation_token_address(&self) -> Address

Returns the address of Reputation Token contract.

pub fn init( &mut self, variable_repository: Address, reputation_token: Address, kyc_token: Address, va_token: Address )

Constructor function.

Note

Initializes contract elements:

Sets up the contract by writing addresses of Variable Repository, Reputation Token, VA Token, KYC Token.

Sets [caller] as the owner of the contract.

Adds [caller] to the whitelist.

Events

OwnerChanged,

AddedToWhitelist,