Module dao::bid_escrow::events

BidEscrow-related events.

Structs

BidCancelled

Event emitted when that a Bid has been cancelled.

BidEscrowSlashResults

Events out the result of the slash in the BidEscrow contract.

BidEscrowVotingCreated

Event emitted when a new VotingStateMachine has been created.

BidSubmitted

Event emitted when a new Bid has been placed.

CSPRTransfer

Event emitted when that a transfer has been made.

JobCancelled

Event emitted when the Job has been canceled.

JobCreated

Event emitted when a new Job has been created.

JobDone

Event emitted when Voting on the Job passed.

JobOfferCreated

Event emitted when a new Job Offer has been created.

JobRejected

Event emitted when Voting on the Job failed.

JobSubmitted

Event emitted when the Job proof has been submitted by the Worker.

Enums

TransferReason

Enum covering all reasons for a transfer in BidEscrow - used in the CSPRTransfer event.