Struct dao::voting::voting_engine::voting_state_machine::VotingStateMachine

pub struct VotingStateMachine { /* private fields */ }

Serializable voting state with a state machine capabilities.

Stores voting metadata, the configuration and the voting progress (stakes).

Implementations

impl VotingStateMachine

pub fn new(voting_id: VotingId, created_at: u64, creator: Address) -> Self

Creates new Voting with immutable Configuration.

pub fn complete_informal_voting( &mut self, configuration: &Configuration ) -> bool

Ends the informal phase, verifies if the result is close, updates voting type to Formal.

pub fn finish(&mut self)

Ends the voting process gracefully.

pub fn cancel(&mut self)

Ends the voting process forcefully and cancels the result.

pub fn voting_type(&self) -> VotingType

Returns the type of Voting.

pub fn is_informal_without_stake(&self, configuration: &Configuration) -> bool

Checks if Voting is of type Informal and stakes the reputation.

pub fn is_in_time(&self, block_time: u64, configuration: &Configuration) -> bool

Checks if Voting is still in the voting phase.

pub fn informal_voting_end_time( &self, configuration: &Configuration ) -> BlockTime

Gets the informal phase end time.

pub fn time_between_votings_end_time( &self, configuration: &Configuration ) -> BlockTime

Gets the informal-formal break end time.

pub fn formal_voting_end_time(&self, configuration: &Configuration) -> BlockTime

Gets the informal phase end time.

pub fn is_in_favor(&self) -> bool

Checks is the in_favor stake surpasses the against stake.

pub fn get_winning_stake(&self) -> Balance

Depending on the result of the voting, returns the amount of reputation staked on the winning side.

pub fn get_result( &self, voters_number: u32, configuration: &Configuration ) -> VotingResult

Gets the current voting result.

pub fn add_stake(&mut self, stake: Balance, choice: Choice)

Adds the stake to the total bound stake.

pub fn add_unbound_stake(&mut self, stake: Balance, choice: Choice)

Adds the stake to the total unbound stake.

pub fn remove_stake(&mut self, stake: Balance, choice: Choice)

Removes the stake from the total bound stake.

pub fn remove_unbound_stake(&mut self, stake: Balance, choice: Choice)

Removes the stake from the total unbound stake.

pub fn bind_stake(&mut self, stake: Balance, choice: Choice)

Removes the unbound stake and adds it to the bound stake.

pub fn total_stake(&self) -> Balance

Gets the sum of bound and unbound stake.

pub fn total_bound_stake(&self) -> Balance

Gets the total bound stake.

pub fn total_unbound_stake(&self) -> Balance

Gets the total unbound stake.

pub fn voting_id(&self) -> VotingId

Get the voting’s voting id.

pub fn stake_in_favor(&self) -> Balance

Get the voting’s stake in favor.

pub fn stake_against(&self) -> Balance

Get the voting’s stake against.

pub fn creator(&self) -> &Address

Gets the voting creator.

pub fn state(&self) -> &VotingState

Gets the current voting state.

pub fn completed(&self) -> bool

Indicates if Voting is finished or canceled.

pub fn state_in_time( &self, block_time: BlockTime, configuration: &Configuration ) -> VotingState

Returns the voting state depending on a given block_time.

pub fn informal_stats(&self) -> &Stats

Gets the Informal Voting statistics.

pub fn formal_stats(&self) -> &Stats

Gets the Formal Voting statistics.

pub fn created_at(&self) -> u64

pub fn guard_vote(&self, block_time: BlockTime, configuration: &Configuration)

Verifies if a ballot can be casted.

Stops contract execution if validation fails. See VoteInTime.

pub fn guard_finish_formal_voting( &self, block_time: BlockTime, configuration: &Configuration )

Verifies if the formal voting can be finished.

Stops contract execution if validation fails. See AfterFormalVoting and VotingNotCompleted.

pub fn guard_cancel_finished_voting( &self, block_time: BlockTime, configuration: &Configuration )