Module dao::voting_contracts

Module containg voting contracts.

Structs

AdminContract

Admin contract uses VotingEngine to vote on changes of ownership and managing whitelists of other contracts.

AdminContractDeployer

Deployer for the AdminContract contract.

AdminContractRef

Reference to the AdminContract contract instance.

KycVoterContract

KycVoterContract

KycVoterContractDeployer

Deployer for the KycVoterContract contract.

KycVoterContractRef

Reference to the KycVoterContract contract instance.

OnboardingRequestContract

Onboarding Request Contract.

OnboardingRequestContractDeployer

Deployer for the OnboardingRequestContract contract.

OnboardingRequestContractRef

Reference to the OnboardingRequestContract contract instance.

OnboardingVotingCreated

Event emitted when onboarding voting has been created.

RepoVoterContract

RepoVoterContract

RepoVoterContractDeployer

Deployer for the RepoVoterContract contract.

RepoVoterContractRef

Reference to the RepoVoterContract contract instance.

ReputationVoterContract

ReputationVoterContract

ReputationVoterContractDeployer

Deployer for the ReputationVoterContract contract.

ReputationVoterContractRef

Reference to the ReputationVoterContract contract instance.

SimpleVoterContract

SimpleVoterContract

SimpleVoterContractDeployer

Deployer for the SimpleVoterContract contract.

SimpleVoterContractRef

Reference to the SimpleVoterContract contract instance.

SlashedVotings

SlashingVoterContract

Slashing Voter contract uses VotingEngine to vote on changes of ownership and managing whitelists of other contracts.

SlashingVoterContractDeployer

Deployer for the SlashingVoterContract contract.

SlashingVoterContractRef

Reference to the SlashingVoterContract contract instance.

Enums

AdminAction

Enum for actions that AdminContract can perform

ReputationAction

Action to perform against reputation